# openapi.yaml
openapi: '3.1.0'
info:
  title: Photo Gallery API
  description: |
    A RESTful API for managing a personal photo gallery.
    
    ## Authentication
    All API endpoints require an API key passed as a Bearer token in the Authorization header:
    ```
    Authorization: Bearer <your-api-key>
    ```
    
    Generate an API key using the CLI script:
    ```
    node scripts/generate-api-key.js --label "My App"
    ```
    
    ## Rate Limiting
    API requests are rate-limited to **100 requests per minute** per API key.
    Rate limit information is returned in response headers:
    - `RateLimit-Limit`: Maximum requests allowed
    - `RateLimit-Remaining`: Remaining requests in the current window
    - `RateLimit-Reset`: Time when the rate limit window resets
    
  version: '1.0.0'
  contact:
    name: Photo Gallery API Support
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://your-domain.com
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: API key generated via `node scripts/generate-api-key.js`
  schemas:
...
    Image:
      type: object
      properties:
        id:
          type: integer
          description: Unique image identifier
          example: 42
        originalName:
          type: string
          description: Original filename of the uploaded image
          example: "sunset-beach.jpg"
        filename:
          type: string
          description: Stored filename on disk
          example: "1720000000000-123456789.jpg"
        mimeType:
          type: string
          description: MIME type of the image
          example: "image/jpeg"
        fileSize:
          type: integer
          description: File size in bytes
          example: 2048576
        width:
          type: integer
          description: Image width in pixels
          example: 3840
        height:
          type: integer
          description: Image height in pixels
          example: 2160
        title:
          type: string
          nullable: true
          description: Optional title
          example: "Sunset at the beach"
        description:
          type: string
          nullable: true
          description: Optional description
          example: "A beautiful sunset photo"
        altText:
          type: string
          nullable: true
          description: Alt text for accessibility
          example: "Orange sunset over the ocean"
        cameraMake:
          type: string
          nullable: true
          description: Camera manufacturer from EXIF
          example: "Canon"
        cameraModel:
          type: string
          nullable: true
          description: Camera model from EXIF
          example: "Canon EOS R5"
        focalLength:
          type: number
          nullable: true
          description: Focal length in mm
          example: 85
        aperture:
          type: number
          nullable: true
          description: Aperture (f-stop)
          example: 1.8
        shutterSpeed:
          type: string
          nullable: true
          description: Shutter speed as fraction string
          example: "1/250"
        iso:
          type: integer
          nullable: true
          description: ISO sensitivity
          example: 400
        takenAt:
          type: string
          format: date-time
          nullable: true
          description: Date/time photo was taken (from EXIF)
        latitude:
          type: number
          nullable: true
          description: GPS latitude
          example: 48.8566
        longitude:
          type: number
          nullable: true
          description: GPS longitude
          example: 2.3522
        isPublic:
          type: boolean
          description: Whether image is publicly accessible
          example: true
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
          description: Associated tags
        createdAt:
          type: string
          format: date-time
          description: Upload timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
...
    Tag:
      type: object
      properties:
        id:
          type: integer
          description: Unique tag identifier
          example: 7
        name:
          type: string
          description: Tag name (lowercase)
          example: "nature"
        createdAt:
          type: string
          format: date-time
    TagCreate:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Tag name (will be lowercased)
          example: "landscape"
    TagUpdate:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: New tag name (will be lowercased)
          example: "landscapes"

    # ─── Pagination ──────────────────────────────────────────────────────────────
    Pagination:
      type: object
      properties:
        page:
          type: integer
          description: Current page number (1-based)
          example: 1
        limit:
          type: integer
          description: Items per page
          example: 20
        total:
          type: integer
          description: Total number of items
          example: 150
        totalPages:
          type: integer
          description: Total number of pages
          example: 8
        hasNext:
          type: boolean
          description: Whether a next page exists
          example: true
        hasPrev:
          type: boolean
          description: Whether a previous page exists
          example: false

    # ─── Errors ──────────────────────────────────────────────────────────────────
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code
              enum:
                - BAD_REQUEST
                - UNAUTHORIZED
                - FORBIDDEN
                - NOT_FOUND
                - CONFLICT
                - RATE_LIMIT_EXCEEDED
                - INTERNAL_ERROR
              example: "NOT_FOUND"
            message:
              type: string
              description: Human-readable error message
              example: "Image with ID 42 not found"

    # ─── Search ──────────────────────────────────────────────────────────────────
    SearchFacets:
      type: object
      properties:
        tags:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
              imageCount:
                type: integer
        cameraMakes:
          type: array
          items:
            type: object
            properties:
              cameraMake:
                type: string
              count:
                type: integer
        cameraModels:
          type: array
          items:
            type: object
            properties:
              cameraModel:
                type: string
              count:
                type: integer

    # ─── Admin ───────────────────────────────────────────────────────────────────
...
security:
  - BearerAuth: []

# ─── Paths ────────────────────────────────────────────────────────────────────
paths:
...
  /api/v1/tags:
    get:
      operationId: listTags
      summary: List tags
      description: Returns a paginated list of all tags.
      tags: [Tags]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: search
          in: query
          schema:
            type: string
          description: Filter tags by name
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Tag'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      operationId: createTag
      summary: Create a tag
      description: Create a new tag. Tag names are automatically lowercased.
      tags: [Tags]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagCreate'
      responses:
        '201':
          description: Tag created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Tag'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Tag already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/tags/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
        description: Tag ID

    get:
      operationId: getTag
      summary: Get a tag
      description: Get a single tag by ID, including a sample of its images.
      tags: [Tags]
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Tag'
        '404':
          description: Tag not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      operationId: updateTag
      summary: Rename a tag
      description: Rename an existing tag. The new name will be lowercased.
      tags: [Tags]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagUpdate'
      responses:
        '200':
          description: Tag updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Tag'
        '404':
          description: Tag not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Tag name conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      operationId: deleteTag
      summary: Delete a tag
      description: Delete a tag and remove it from all images.
      tags: [Tags]
      responses:
        '204':
          description: Tag deleted successfully
        '404':
          description: Tag not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── Search ──────────────────────────────────────────────────────────────────
  /api/v1/search:
    get:
      operationId: searchImages
      summary: Search images
      description: Full-text search with multiple filter options.
      tags: [Search]
      parameters:
        - name: q
          in: query
          schema:
            type: string
          description: Search query (searches filename, title, description)
        - name: tags
          in: query
          schema:
            type: string
          description: Comma-separated list of tag names to filter by
          example: "nature,sunset"
        - name: cameraMake
          in: query
          schema:
            type: string
          description: Filter by camera manufacturer
        - name: cameraModel
          in: query
          schema:
            type: string
          description: Filter by camera model
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
          description: Filter images uploaded on or after this date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
          description: Filter images uploaded on or before this date
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [createdAt, updatedAt, fileSize, originalName]
            default: createdAt
        - name: sortDir
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Image'
                  query:
                    type: object
                    description: Parsed query parameters
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/v1/search/facets:
    get:
      operationId: getSearchFacets
      summary: Get search facets
      description: Returns aggregated facet data for building filter UI (top tags, camera makes, models).
      tags: [Search]
      responses:
        '200':
          description: Facet data
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/SearchFacets'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ── Admin ────────────────────────────────────────────────────────────────────
...